add tableReader to implement tupleReader, make physicalJoinOperator abstract, add time print
