+++ src/main/java/model/TupleReader.java
@@ -1,5 +1,9 @@
 package model;
 
+import java.io.IOException;
+
 public interface TupleReader{
+    void init();
     Tuple readNextTuple();
+    void readPage();
 }
